refactor:" changed  export structure and styles"
