+++ src/contexts/index.js
@@ -1,0 +1,1 @@
+export { useFilter, FilterProvider } from "./filterContext/context";
